+++ packages/terra-grid/src/terra-dev-site/test/grid/GridNested.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Grid from '../../../../lib/Grid';
-import './GridExample.scss';
+import './GridExample.module.scss';
 
 const GridNested = () => (
   <Grid id="grid">